Fix java.lang.UnsatisfiedLinkError
Occurs when program runs on Macs with the following error:
Loading library prism_es2 from resource failed: java.lang.UnsatisfiedLinkError: Can't load library
